data refactor
