+++ src/pages/api/account/register.ts
@@ -1,6 +1,7 @@
 import { registerSchema } from '@/features/account/accountTypes';
+import { query } from '@/data/db';
+import { User } from '@/data/schema';
 import { createApiRouter, onError } from '@/utils/router';
-import { User, query } from '@/utils/db';
 import { hashPassword } from '@/utils/hash';
 
 const router = createApiRouter();
