+++ src/pages/api/manage/change-password.ts
@@ -1,6 +1,7 @@
 import { changePasswordSchema } from '@/features/manage/manageTypes';
+import { query } from '@/data/db';
+import { User } from '@/data/schema';
 import { createApiRouter, onError, authorize } from '@/utils/router';
-import { User, query } from '@/utils/db';
 import { hashPassword, verifyPassword } from '@/utils/hash';
 
 const router = createApiRouter();
