+++ src/pages/api/manage/index.ts
@@ -2,8 +2,9 @@ import {
     deleteAccountSchema,
     updateProfileSchema
 } from '@/features/manage/manageTypes';
+import { query } from '@/data/db';
+import { User } from '@/data/schema';
 import { createApiRouter, onError, authorize } from '@/utils/router';
-import { User, query } from '@/utils/db';
 
 const router = createApiRouter();
